// ui/src/api/generated/model/activeDeviceResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Whirlpool PDI Tool API
 * Whirlpool PDI Tool API APIs developed by Scopt Analytics.
 * OpenAPI spec version: 1.7.1
 */

export interface ActiveDeviceResponse {
  uuid: string;
  imei: string;
  device_type: string;
  device_fingerprint: string;
  display_label: string;
  is_current?: boolean;
  has_active_session?: boolean;
  last_seen_at?: string | null;
}